requirements.txt
''' pip freeze --- команды выводящяя список используемых пакетов '''

alembic==1.7.5
aniso8601==9.0.1
astroid==2.9.0
bcrypt==3.2.0
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
coverage==6.2
coveralls==3.3.1
cryptography==36.0.1
dnspython==2.1.0
docopt==0.6.2
email-validator==1.1.3
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
gunicorn==20.1.0
httpie==2.1.0
idna==3.3
importlib-metadata==4.10.0
importlib-resources==5.4.0
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
lazy-object-proxy==1.6.0
Mako==1.1.6
MarkupSafe==2.0.1
mccabe==0.6.1
mysql-connector==2.2.9
mysql-connector-python==8.0.27
platformdirs==2.4.0
protobuf==3.19.1
psycopg2-binary==2.9.2
pycparser==2.21
Pygments==2.7.4
pylint==2.12.2
PyMySQL==1.0.2
python-dateutil==2.8.2
python-decouple==3.5
python-dotenv==0.19.2
python-editor==1.0.4
pytz==2021.3
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.27
SQLAlchemy-Utils==0.37.9
toml==0.10.2
typing-extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
WTForms==3.0.0
WTForms-SQLAlchemy==0.3
zipp==3.6.0
